Address info pane review cleanup
Reuse shared info pane types, trim speculative marker caching, and simplify copied-state and trajectory filter logic after review.
